docs: added Deprecated comments in front of APIs moved to Compiler interface
